Update to Solr 5.3.1, and some fixes
Update to Solr 5.3.1
Fix path ordering issue in the Dockerfile
In update.sh, remove KEYS file after use

+++ update.sh
@@ -73,4 +73,5 @@ for version in "${versions[@]}"; do
 	)
 done
 
-rm "$upstream_versions"
+if [ -f KEYS ]; then rm KEYS; fi
+if [ -f "$upstream_versions" ]; then rm "$upstream_versions"; fi
